+++ src/kbmod/search/stack_search.h
@@ -30,11 +30,11 @@ using Image = search::Image;
 class StackSearch {
 public:
     StackSearch(ImageStack& imstack);
-    int compute_max_results();
+    uint64_t compute_max_results();
     int num_images() const { return stack.img_count(); }
     int get_image_width() const { return stack.get_width(); }
     int get_image_height() const { return stack.get_height(); }
-    int get_image_npixels() const { return stack.get_npixels(); }
+    uint64_t get_image_npixels() const { return stack.get_npixels(); }
     const ImageStack& get_imagestack() const { return stack; }
 
     // Parameter setters used to control the searches.
@@ -48,7 +48,7 @@ public:
 
     // The primary search functions
     void evaluate_single_trajectory(Trajectory& trj);
-    Trajectory search_linear_trajectory(short x, short y, float vx, float vy);
+    Trajectory search_linear_trajectory(int x, int y, float vx, float vy);
     void prepare_search(std::vector<Trajectory>& search_list, int min_observations);
     std::vector<Trajectory> search_single_batch();
     void search_batch();
